added check for disallowed games in get_overlap
In the original code, the method Person.get_overlap does not check if a game is completely disallowed, and only returns the best match between two players.  If the 'best' match for two players is a disallowed game, this will result in not adding an edge between these two players, even if there is a valid match for other games.  This commit fixes this by checking for disallowed games in Person.get_overlap() instead of checking later.
values.txt has been updated to include a simple example of when this is relevant